requirements.txt
requests
pytesseract
pyautogui
keyboard
opencv-python
numpy
Pillow
PyQt5
python-dotenv